+++ src/kls_region.h
@@ -159,7 +159,7 @@ void KLS_autoregion_on_temp_free(struct Koliseo_Temp* t_kls);
 void KLS_autoregion_on_temp_push(struct Koliseo_Temp* t_kls, ptrdiff_t padding, const char* caller, void* user);
 
 #ifndef KLS_DEFAULT_HOOKS
-#define KLS_DEFAULT_HOOKS &(KLS_Hooks) { \
+#define KLS_DEFAULT_HOOKS (KLS_Hooks) { \
         .on_new_handler = &KLS_autoregion_on_new, \
         .on_free_handler = &KLS_autoregion_on_free, \
         .on_push_handler =  &KLS_autoregion_on_push, \
@@ -168,20 +168,6 @@ void KLS_autoregion_on_temp_push(struct Koliseo_Temp* t_kls, ptrdiff_t padding, 
         .on_temp_push_handler = &KLS_autoregion_on_temp_push, \
     }
 #endif // KLS_DEFAULT_HOOKS
-
-/**
- * DEPRECATED: Support for multiple extension will be dropped in the next release.
- */
-#ifndef KLS_DEFAULT_EXTENSIONS_LEN
-#define KLS_DEFAULT_EXTENSIONS_LEN 1
-#endif // KLS_DEFAULT_EXTENSIONS_LEN
-
-/**
- * DEPRECATED: Support for multiple extension will be dropped in the next release.
- */
-#ifndef KLS_AUTOREGION_EXT_SLOT
-#define KLS_AUTOREGION_EXT_SLOT 0
-#endif // KLS_AUTOREGION_EXT_SLOT
 
 #include "koliseo.h"
 
@@ -226,16 +212,16 @@ void *kls_temp_push_zero_typed_dbg(Koliseo_Temp * t_kls, ptrdiff_t size,
 #endif // KOLISEO_HAS_LOCATE
 
 #undef KLS_PUSH_ARR_NAMED
-#define KLS_PUSH_ARR_NAMED(kls, type, count, name, desc) (type*)kls_push_zero_named((kls), sizeof(type), _Alignof(type), (count), (name), (desc))
+#define KLS_PUSH_ARR_NAMED(kls, type, count, name, desc) (type*)kls_push_zero_named((kls), sizeof(type), KLS_ALIGNOF(type), (count), (name), (desc))
 
 #undef KLS_PUSH_ARR_TYPED
-#define KLS_PUSH_ARR_TYPED(kls, type, count, region_type, name, desc) (type*)kls_push_zero_typed((kls), sizeof(type), _Alignof(type), (count), (region_type), (name), (desc))
+#define KLS_PUSH_ARR_TYPED(kls, type, count, region_type, name, desc) (type*)kls_push_zero_typed((kls), sizeof(type), KLS_ALIGNOF(type), (count), (region_type), (name), (desc))
 
 #undef KLS_PUSH_ARR_T_NAMED
-#define KLS_PUSH_ARR_T_NAMED(kls_temp, type, count, name, desc) (type*)kls_temp_push_zero_named((kls_temp), sizeof(type), _Alignof(type), (count), (name), (desc))
+#define KLS_PUSH_ARR_T_NAMED(kls_temp, type, count, name, desc) (type*)kls_temp_push_zero_named((kls_temp), sizeof(type), KLS_ALIGNOF(type), (count), (name), (desc))
 
 #undef KLS_PUSH_ARR_T_TYPED
-#define KLS_PUSH_ARR_T_TYPED(kls_temp, type, count, region_type, name, desc) (type*)kls_temp_push_zero_typed((kls_temp), sizeof(type), _Alignof(type), (count), (region_type), (name), (desc))
+#define KLS_PUSH_ARR_T_TYPED(kls_temp, type, count, region_type, name, desc) (type*)kls_temp_push_zero_typed((kls_temp), sizeof(type), KLS_ALIGNOF(type), (count), (region_type), (name), (desc))
 
 KLS_Region_List kls_rl_emptyList(void);
 #define KLS_RL_GETLIST() kls_rl_emptyList()