+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [0.6.0] - Unreleased
+
+### Added
+
+- Define fallback for `KLS_ALIGNOF` for `<C11`
+  - Allows building for older standards
+
+### Changed
+
+- Drop support for multiple extensions
+- Drop deprecated symbols
+  - `kls__check_available()` macro -> now a function (previously named `kls__check_available_failable()`)
+  - `KLS_DEFAULT_EXTENSIONS_LEN`, `KLS_MAX_EXTENSIONS`
+  - `KLS_AUTOREGION_EXT_SLOT` (from `src/kls_region.[h,c]`)
+- Drop `kls_pop_AR()`, `kls_temp_pop_AR()`
+- Refactor `kls_clear()`:
+  - Frees the chained `Koliseo`s
+  - Calls `on_new_handler()` to reinit the extension
+- Since extension data is shared in the `Koliseo` chain:
+  - Made `on_free_handler()` use the extension data properly
+  - Made `kls__try_grow()` pass `NULL` `on_new_handler` to the next `Koliseo`
+  - Made `kls_free()` clear up a pending `Koliseo_Temp` before calling `on_free()`
+- Drop `kls_` prefix from `KLS_Conf` fields
+- Fix direct usage of `_Alignof` in `kls_region.h`
+- Move `CODEOWNERS`, `SECURITY.md` to `.github/`
+- Move `format.sh`, `installer-sh`, `.astylerc` to `scripts/`
+- Bump `invil` to `0.2.29`
+- Bump `amboso` to `2.1.1`
+
 ## [0.5.10] - 2026-01-10
 
 ### Added